Add files via upload
Deposit and withdraw buttons are working, but you have to hard code the user id (This can be done on line 16 in deposit.inc.php, or on line 5 in withdraw.inc.php by replacing the '?' with the integer value that corresponds with the users idUsers)...this is a bug. I am in the process of debugging.